pyproject.toml
[tool.poetry]
name = "quack-project"
version = "0.1.0"
description = ""
authors = ["Jay <joaomarcelo1015ja@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.7"
pillow = "^10.4.0"
factory-boy = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
